pc - WIP full test coverage from vitest
