+++ vite-project/src/main/layouts/BasicLayout/BasicLayout.jsx
@@ -1,7 +1,9 @@
 import { Container } from "react-bootstrap";
 import Footer from "main/components/Nav/Footer";
 import AppNavbar from "main/components/Nav/AppNavbar";
-import { useCurrentUser, useLogout } from "main/utils/currentUser";
+import { useCurrentUser } from "main/utils/useCurrentUser";
+import { useLogout } from "main/utils/useLogout";
+
 import { useSystemInfo } from "main/utils/systemInfo";
 
 export default function BasicLayout({ children }) {
