+++ vite-project/src/main/utils/useCurrentUser.js
@@ -1,6 +1,5 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery} from "@tanstack/react-query";
 import axios from "axios";
-import { useNavigate } from "react-router";
 
 export function useCurrentUser() {
   const queryResults = useQuery({
@@ -29,19 +28,6 @@ export function useCurrentUser() {
     initialData: { loggedIn: false, root: null, initialData: true },
   });
   return queryResults.data;
-}
-
-export function useLogout() {
-  const queryClient = useQueryClient();
-  const navigate = useNavigate();
-  const mutation = useMutation({
-    mutationFn: async () => {
-      await axios.post("/logout");
-      await queryClient.resetQueries({ queryKey: ["current user"] });
-      navigate("/");
-    },
-  });
-  return mutation;
 }
 
 export function hasRole(currentUser, role) {
