+++ vite-project/src/tests/components/Nav/Footer.test.jsx
