--- vite-project/src/tests/utils/currentUser.test.jsx
@@ -1,277 +1,0 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { useCurrentUser, useLogout, hasRole } from "main/utils/currentUser";
-import { renderHook, waitFor, act } from "@testing-library/react";
-import {
-  apiCurrentUserFixtures,
-  currentUserFixtures,
-} from "fixtures/currentUserFixtures";
-import mockConsole from "tests/testutils/mockConsole";
-import { useNavigate } from "react-router";
-import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
-
-import axios from "axios";
-import AxiosMockAdapter from "axios-mock-adapter";
-
-vi.mock("react-router", async () => {
-  const actual = await vi.importActual("react-router");
-  return {
-    ...actual,
-    MemoryRouter: actual.MemoryRouter,
-  };
-});
-
-let axiosMock;
-describe("utils/currentUser tests", () => {
-  describe("useCurrentUser tests", () => {
-    beforeEach(() => {
-      axiosMock = new AxiosMockAdapter(axios);
-    });
-
-    test("useCurrentUser retrieves initial data", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          {children}
-        </QueryClientProvider>
-      );
-
-      axiosMock.onGet("/api/currentUser").timeoutOnce();
-      axiosMock
-        .onGet("/api/systemInfo")
-        .reply(200, systemInfoFixtures.showingNeither);
-
-      const restoreConsole = mockConsole();
-
-      const { result } = renderHook(() => useCurrentUser(), { wrapper });
-      await waitFor(() =>
-        expect(queryClient.getQueryState(["current user"]).status).toBe(
-          "success",
-        ),
-      );
-
-      expect(result.current).toEqual({
-        loggedIn: false,
-        root: null,
-        initialData: true,
-      });
-
-      const queryState = queryClient.getQueryState(["current user"]);
-      expect(queryState).toBeDefined();
-
-      queryClient.clear();
-
-      await waitFor(() => expect(console.error).toHaveBeenCalled());
-      const errorMessage = console.error.mock.calls[0][0];
-      expect(errorMessage).toMatch(/Error invoking axios.get:/);
-      restoreConsole();
-    });
-
-    test("useCurrentUser retrieves data from API", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          {children}
-        </QueryClientProvider>
-      );
-
-      const axiosMock = new AxiosMockAdapter(axios);
-      axiosMock
-        .onGet("/api/currentUser")
-        .reply(200, apiCurrentUserFixtures.userOnly);
-      axiosMock
-        .onGet("/api/systemInfo")
-        .reply(200, systemInfoFixtures.showingNeither);
-
-      const { result } = renderHook(() => useCurrentUser(), { wrapper });
-
-      await waitFor(() =>
-        expect(
-          queryClient.getQueryState(["current user"]).dataUpdateCount,
-        ).toBe(1),
-      );
-
-      expect(result.current).toEqual(currentUserFixtures.userOnly);
-      queryClient.clear();
-    });
-
-    test("useCurrentUser when API unreachable", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          {children}
-        </QueryClientProvider>
-      );
-
-      const axiosMock = new AxiosMockAdapter(axios);
-      axiosMock.onGet("/api/currentUser").reply(404);
-
-      const restoreConsole = mockConsole();
-      const { result } = renderHook(() => useCurrentUser(), { wrapper });
-
-      await waitFor(() =>
-        expect(queryClient.getQueryState(["current user"]).status).toBe(
-          "success",
-        ),
-      );
-      expect(console.error).toHaveBeenCalled();
-      const errorMessage = console.error.mock.calls[0][0];
-      expect(errorMessage).toMatch(/Error invoking axios.get:/);
-      restoreConsole();
-
-      expect(result.current).toEqual({
-        initialData: true,
-        loggedIn: false,
-        root: null,
-      });
-      queryClient.clear();
-    });
-
-    test("useCurrentUser when API returns 403 (not logged in)", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          {children}
-        </QueryClientProvider>
-      );
-
-      const axiosMock = new AxiosMockAdapter(axios);
-      axiosMock.onGet("/api/currentUser").reply(403);
-
-      const restoreConsole = mockConsole();
-      const { result } = renderHook(() => useCurrentUser(), { wrapper });
-
-      await waitFor(() =>
-        expect(
-          queryClient.getQueryState(["current user"]).dataUpdateCount,
-        ).toBe(1),
-      );
-      expect(console.error).not.toHaveBeenCalled();
-      restoreConsole();
-
-      expect(result.current).toEqual({
-        loggedIn: false,
-        root: {},
-      });
-      queryClient.clear();
-    });
-
-    test("useCurrentUser handles missing roles correctly", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          {children}
-        </QueryClientProvider>
-      );
-
-      const apiResult = apiCurrentUserFixtures.missingRolesToTestErrorHandling;
-      const axiosMock = new AxiosMockAdapter(axios);
-      axiosMock.onGet("/api/currentUser").reply(200, apiResult);
-
-      const restoreConsole = mockConsole();
-      const { result } = renderHook(() => useCurrentUser(), { wrapper });
-
-      await waitFor(() =>
-        expect(
-          queryClient.getQueryState(["current user"]).dataUpdateCount,
-        ).toBe(1),
-      );
-      expect(console.error).toHaveBeenCalled();
-      const errorMessage = console.error.mock.calls[0][0];
-      expect(errorMessage).toMatch(/Error getting roles: /);
-      restoreConsole();
-
-      let expectedResult = {
-        loggedIn: true,
-        root: { ...apiResult, rolesList: ["ERROR_GETTING_ROLES"] },
-      };
-      expect(result.current).toEqual(expectedResult);
-      queryClient.clear();
-    });
-  });
-  describe("useLogout tests", () => {
-    beforeEach(() => {
-      axiosMock = new AxiosMockAdapter(axios);
-    });
-    test("useLogout", async () => {
-      const queryClient = new QueryClient();
-      const wrapper = ({ children }) => (
-        <QueryClientProvider client={queryClient}>
-          <MemoryRouter>{children}</MemoryRouter>
-        </QueryClientProvider>
-      );
-
-      const axiosMock = new AxiosMockAdapter(axios);
-      axiosMock.onPost("/logout").reply(200);
-
-      const navigateSpy = vi.fn();
-      useNavigate.mockImplementation(() => navigateSpy);
-
-      const resetQueriesSpy = jest.spyOn(queryClient, "resetQueries");
-
-      const { result } = renderHook(() => useLogout(), { wrapper });
-
-      act(() => {
-        expect(useNavigate).toHaveBeenCalled();
-        result.current.mutate();
-      });
-      await waitFor(() => expect(navigateSpy).toHaveBeenCalledWith("/"));
-
-      await waitFor(() =>
-        expect(resetQueriesSpy).toHaveBeenCalledWith({
-          queryKey: ["current user"],
-        }),
-      );
-
-      queryClient.clear();
-    });
-  });
-  describe("hasRole tests", () => {
-    test('hasRole(x,"ROLE_ADMIN") return falsy when currentUser ill-defined', async () => {
-      expect(hasRole(null, "ROLE_ADMIN")).toBeFalsy();
-      expect(hasRole({}, "ROLE_ADMIN")).toBeFalsy();
-      expect(hasRole({ loggedIn: null }, "ROLE_ADMIN")).toBeFalsy();
-      expect(hasRole({ loggedIn: true }, "ROLE_ADMIN")).toBeFalsy();
-      expect(hasRole({ loggedIn: true, root: null }, "ROLE_ADMIN")).toBeFalsy();
-      expect(hasRole({ loggedIn: true, root: {} }, "ROLE_ADMIN")).toBeFalsy();
-      expect(
-        hasRole({ loggedIn: true, root: { rolesList: null } }, "ROLE_ADMIN"),
-      ).toBeFalsy();
-    });
-
-    test('hasRole(x,"ROLE_ADMIN") returns correct values when currentUser properly defined', async () => {
-      expect(
-        hasRole({ loggedIn: true, root: { rolesList: [] } }, "ROLE_ADMIN"),
-      ).toBeFalsy();
-      expect(
-        hasRole(
-          { loggedIn: true, root: { rolesList: ["ROLE_USER"] } },
-          "ROLE_ADMIN",
-        ),
-      ).toBeFalsy();
-      expect(
-        hasRole(
-          { loggedIn: true, root: { rolesList: ["ROLE_USER", "ROLE_ADMIN"] } },
-          "ROLE_ADMIN",
-        ),
-      ).toBeTruthy();
-    });
-
-    test('hasRole(x,"ROLE_USER") returns correct values when currentUser properly defined', async () => {
-      expect(
-        hasRole({ loggedIn: true, root: { rolesList: [] } }, "ROLE_USER"),
-      ).toBeFalsy();
-      expect(
-        hasRole(
-          { loggedIn: true, root: { rolesList: ["ROLE_USER"] } },
-          "ROLE_USER",
-        ),
-      ).toBeTruthy();
-      expect(
-        hasRole(
-          { loggedIn: true, root: { rolesList: ["ROLE_USER", "ROLE_ADMIN"] } },
-          "ROLE_USER",
-        ),
-      ).toBeTruthy();
-    });
-  });
-});
