+++ vite-project/src/tests/utils/useLogout.test.jsx
@@ -1,0 +1,71 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import {  useLogout } from "main/utils/useLogout";
+import { renderHook, waitFor, act } from "@testing-library/react";
+import { MemoryRouter } from "react-router";
+import axios from "axios";
+import AxiosMockAdapter from "axios-mock-adapter";
+import { vi } from "vitest";
+
+
+// The mock MUST be at the top level of the file
+const navigateSpy = vi.fn();
+vi.mock('react-router-dom', async (importOriginal) => {
+  const actual = await importOriginal();
+  return {
+    ...actual,
+    useNavigate: () => navigateSpy,
+  };
+});
+
+let axiosMock;
+describe("useLogout tests", () => {
+
+  let queryClient;
+  let axiosMock;
+  let resetQueriesSpy;
+
+  beforeEach(() => {
+    queryClient = new QueryClient();
+    axiosMock = new AxiosMockAdapter(axios);
+    resetQueriesSpy = vi.spyOn(queryClient, 'resetQueries');
+  });
+
+  afterEach(() => {
+    axiosMock.restore();
+    queryClient.clear();
+    vi.restoreAllMocks();
+  });
+
+  test('should log out the user, reset queries, and navigate to home', async () => {
+    axiosMock.onPost('/logout').reply(200);
+
+    const wrapper = ({ children }) => (
+      <QueryClientProvider client={queryClient}>
+        <MemoryRouter>{children}</MemoryRouter>
+      </QueryClientProvider>
+    );
+
+    // Render the hook
+    const { result } = renderHook(() => useLogout(), { wrapper });
+
+    // Use a guard to ensure result.current is not null before proceeding
+    if (!result.current) {
+      console.error("Hook failed to render, result.current is null.");
+      // This assertion will make the test fail explicitly if the hook is null
+      expect(result.current).not.toBeNull();
+      return;
+    }
+
+    // Now it's safe to call mutate
+    act(() => {
+      result.current.mutate();
+    });
+
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+
+    expect(axiosMock.history.post.length).toBe(1);
+    expect(resetQueriesSpy).toHaveBeenCalledWith({ queryKey: ['current user'] });
+  });
+});
+
+
